pyproject.toml
[project]
name = "c-analysis"
version = "0.1.0"
description = "CFPB Wells Fargo consumer complaints — emerging topics, redress LOB taxonomy, risk signals, dashboard."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas>=2.0.0,<3",
    "pyarrow>=14.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    # numpy 2.x is not yet compatible with torch/numba wheels we rely on
    "numpy<2.0.0",
    "streamlit>=1.28.0",
    # Narrative pipeline (embeddings + BERTopic). torch<2.3 = last with macOS x86_64 wheels.
    "torch>=2.0.0,<2.3",
    "bertopic>=0.16.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "ruff>=0.1.0",
    "pytest>=7.4.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.1.0", "pytest>=7.4.0"]

[project.scripts]
run-all = "scripts.run_all:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
